Create Hello
Python is a popular, general-purpose programming language and has straightforward syntax. Python also provides many built-in functions. In this tutorial, we will learn how to start with Python programming.

This is the most basic Python program. It specifies how to print any statement in Python.

In old Python (up to Python 2.7.0), the print command is not written in parenthesis, but in new python software (Python 3.4.3), it is mandatory to add parenthesis to print a statement.

Let' see the following example to print Hello World