Create Find the area of a triangle
Mathematical formula:

Area of a triangle = (s*(s-a)*(s-b)*(s-c))-1/2
Here is the semi-perimeter and a, b and c are three sides of the triangle. Let's understand the following example.

We take the input as three sides of a triangle and store them in three variables a, b, and c. Then, we have calculated the semi-perimeter of the triangle and put this value in the triangle's area formula.