Create Solve quadratic equation using second method using
Second Method
We can get the solution of the quadric equation by using the direct formula. Let's understand the following example.

The above formula consists of the following cases.

In the above code, we have imported the math module and defined the formula to calculate discriminant. We have then defined findRoots function which take three integer values as arguments. Then we have checked the roots using the if-elif-else statement.

If b2 < 4ac, then the roots are complex (not real). For example - x2 + x + 1, roots are -0.5 + i1.73205 and +0.5 - i1.73205.
If b2 == 4ac, then the both roots are same For example - x2 + x + 1, roots are -0.5 + i1.73205 and +0.5 - i1.73205.
If b2 > 4ac, then the roots are real and different. For example - x2 - 7 x - 12, roots are 3 and 4.