Create Swap two variables
In this tutorial, we will learn how to swap two variables in a Python program. Suppose we have two variables, P and Q; We have to write a Python program for swapping their values. We will also discuss the different methods in Python for doing this task.

Method 1: By using Naïve Approach
In this method, the naïve approach will store the value of the P variable in a temporary variable, and then it will assign the variable P with the value of the Q variable. Then, it will assign the value of the temporary variable to the Q variable, which will result in swapping the values of both the variable.