Create Swap two variables by using XOR method
Method 3: By using XOR method
We can also use the bitwise XOR method for swapping two variables. The XOR of two variables, P and Q, will return the number which has all the bits as 1 whenever the bits of the P and Q variables differ.

Such as XOR of 4 (in binary 0100) and 6 (in binary 0110) is 1010.

XOR of 2 (in binary 0010) and 8 (in binary 1000) is 1010.