Create Convert Kilometres to Miles
Python program to convert Kilometres to Miles
In this tutorial, we will learn a step-by-step approach for how to write a Python program to convert kilometres value into miles value.

kilometre_1 = float (input ("Please enter the speed of car in Kilometre as a unit: "))  
conversion_ratio_1 = 0.621371  
miles_1 = kilometre_1 * conversion_ratio_1  
print ("The speed value of car in Miles: ", miles_1)  

Understanding the Logic Behind the Conversion of Kilometre to Miles
Let's begin with the basics; first, we will discuss the measuring units. As we know, Kilometre and Miles represent the unit of length.

1 kilometre equals 0.62137 miles.  
Miles = kilometre * 0.62137  
And,  
Kilometre = Miles / 0.62137  
Hence, the value "0.62137" would be considered as the factor of conversation or the ratio for converting units.